Addressing TODOs
