+++ .github/workflows/continuous-integration.yml
@@ -54,7 +54,6 @@ jobs:
         flavor: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]
         include:
           # Node 12.15
-          # TODO Add comments about why we test 12.15; I think git blame says it's because of an ESM behavioral change that happened at 12.16
           - flavor: 1
             node: 12.15
             nodeFlag: 12_15
